+++ src/index.ts
@@ -1,5 +1,11 @@
 export * as atomic from './atomic';
-export { BaseModel, connectModels, IDeleteOptions, ISaveOptions } from './baseModel';
+export {
+  BaseModel,
+  connectModels,
+  IDeleteOptions,
+  InterfaceForModel,
+  ISaveOptions,
+} from './baseModel';
 export { Container } from './container';
 export { Model } from './model';
 export { Partition } from './partition';