+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.1 - 2020-04-27
+
+- Add `atomic.findOrCreate`
+- Add `model.toObject` and `toJSON()` (which uses toObject by default)
+
 ## 0.1.0 - 2020-04-24
 
 Initial Release